Module field_types created, type text field tested
