+++ src/__init__.py
@@ -1,0 +1,3 @@
+from .field_types import\
+	PdfFieldType,\
+	get_field_type
