+++ src/pypdf2_fields.py
@@ -1,4 +1,3 @@
-from enum import Enum
 from PyPDF2 import PdfFileWriter
 from PyPDF2.generic import BooleanObject,\
 	IndirectObject, NameObject, TextStringObject
@@ -12,19 +11,6 @@ _KEY_NEED_APPEARANCES = "/NeedAppearances"
 _KEY_PARENT = "/Parent"
 _KEY_T = "/T"
 _KEY_V = "/V"
-
-_FIELD_TYPE = "/FT"
-_TYPE_BUTTON = "/Btn"
-_TYPE_TEXT = "/Tx"
-
-
-class PdfFieldType(Enum):
-	"""
-	This enumeration represents the field types that a PDF file can contain.
-	"""
-	TEXT_FIELD = 0
-	CHECKBOX = 1
-	RADIO_BTN_GROUP = 2
 
 
 class RadioBtnGroup:
@@ -116,32 +102,6 @@ class RadioBtnGroup:
 		"""
 		return self._name
 
-
-def get_field_type(pdf_field):
-	"""
-	Determines the type of the given PDF field: text field, checkbox or radio
-	button group.
-
-	Args:
-		pdf_field (dict): a dictionary that represents a field of a PDF file
-
-	Returns:
-		PdfFieldType: the type of pdf_field of None if no type is determined
-	"""
-	type_val = pdf_field.get(_FIELD_TYPE)
-
-	if type_val == _TYPE_TEXT:
-		return PdfFieldType.TEXT_FIELD
-
-	elif type_val == _TYPE_BUTTON:
-		if _KEY_KIDS in pdf_field:
-			return PdfFieldType.RADIO_BTN_GROUP
-
-		else:
-			return PdfFieldType.CHECKBOX
-
-	else:
-		return None
 
 
 def _make_radio_btn_group_dict(radio_btn_groups):
